# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LearnMate AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /exam-plans:
    get:
      operationId: listExamPlans
      tags: [exam-plans]
      summary: List the student's exam plans with their plan limits
      responses:
        "200":
          description: Exam plans and limits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamPlanList"
    post:
      operationId: createExamPlan
      tags: [exam-plans]
      summary: Create an exam plan (AI-generated day-by-day schedule)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExamPlanInput"
      responses:
        "201":
          description: Generated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamPlanDetail"
        "403":
          description: Free plan limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /exam-plans/{planId}:
    get:
      operationId: getExamPlan
      tags: [exam-plans]
      summary: Get one exam plan with its day-by-day schedule
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plan detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamPlanDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteExamPlan
      tags: [exam-plans]
      summary: Delete an exam plan
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /exam-plans/{planId}/regenerate:
    post:
      operationId: regenerateExamPlan
      tags: [exam-plans]
      summary: Regenerate the plan's schedule (Premium)
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegenerateExamPlanInput"
      responses:
        "200":
          description: Regenerated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamPlanDetail"
        "403":
          description: Premium required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /exam-plans/{planId}/days/{dayId}:
    patch:
      operationId: updateExamPlanDay
      tags: [exam-plans]
      summary: Mark a plan day complete or not complete
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
        - name: dayId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExamPlanDayInput"
      responses:
        "200":
          description: Updated day and plan progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamPlanDayUpdateResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ExamPlanSummary:
      type: object
      properties:
        id:
          type: integer
        batch:
          type: string
          enum: [gcse, a-level, ssc, hsc]
        title:
          type: string
        examDate:
          type: string
          format: date
        subjects:
          type: array
          items:
            type: string
        weakTopics:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, archived]
        dayCount:
          type: integer
        completedDayCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - batch
        - title
        - examDate
        - subjects
        - weakTopics
        - status
        - dayCount
        - completedDayCount
        - createdAt
    ExamPlanDay:
      type: object
      properties:
        id:
          type: integer
        dayIndex:
          type: integer
        date:
          type: string
          format: date
        subject:
          type: string
        topic:
          type: string
        focus:
          type: string
        explanation:
          type: string
        completed:
          type: boolean
      required: [id, dayIndex, date, subject, topic, focus, explanation, completed]
    ExamPlanLimits:
      type: object
      properties:
        plan:
          type: string
          enum: [free, premium, owner]
        maxActivePlans:
          type: ["integer", "null"]
          description: Max active plans; null means unlimited
        maxHorizonDays:
          type: integer
        canCreate:
          type: boolean
        canRegenerate:
          type: boolean
      required: [plan, maxActivePlans, maxHorizonDays, canCreate, canRegenerate]
    ExamPlanList:
      type: object
      properties:
        plans:
          type: array
          items:
            $ref: "#/components/schemas/ExamPlanSummary"
        limits:
          $ref: "#/components/schemas/ExamPlanLimits"
      required: [plans, limits]
    ExamPlanDetail:
      type: object
      properties:
        plan:
          $ref: "#/components/schemas/ExamPlanSummary"
        days:
          type: array
          items:
            $ref: "#/components/schemas/ExamPlanDay"
      required: [plan, days]
    CreateExamPlanInput:
      type: object
      properties:
        batch:
          type: string
          enum: [gcse, a-level, ssc, hsc]
        subjects:
          type: array
          items:
            type: string
        examDate:
          type: string
          format: date
        weakTopics:
          type: array
          items:
            type: string
        title:
          type: string
      required: [batch, subjects, examDate]
    RegenerateExamPlanInput:
      type: object
      properties:
        adjustment:
          type: string
          description: Optional instruction, e.g. "more time on Physics"
    UpdateExamPlanDayInput:
      type: object
      properties:
        completed:
          type: boolean
      required: [completed]
...
    ExamPlanDayUpdateResult:
      type: object
      properties:
        day:
          $ref: "#/components/schemas/ExamPlanDay"
        dayCount:
          type: integer
        completedDayCount:
          type: integer
      required: [day, dayCount, completedDayCount]
...